+++ lib/buckets/list.js
@@ -8,21 +8,17 @@
  * Copyright (c) 2019, Joyent, Inc.
  */
 
-var common = require('../common');
+var common = require('./common');
 
 function listBuckets(req, res, next) {
 
     var log = req.log;
     log.debug({
-        bucket: req.key
+        key: req.key,
+        params: req.params
     }, 'listBuckets: requested');
 
-    var obj = {
-        bucket_data: req.bucket_data
-    };
-
-    /* This function uses an emitter rather than callback pattern for speed. */
-    var mreq = common.bucketList(req.key, req, obj);
+    var mreq = common.listBuckets(req);
 
     mreq.once('error', function onError(err) {
         mreq.removeAllListeners('end');
@@ -42,7 +38,6 @@ function listBuckets(req, res, next) {
         res.end();
         next();
     });
-
 }
 
 module.exports = {
