+++ lib/buckets/objects/list.js
@@ -8,23 +8,22 @@
  * Copyright (c) 2019, Joyent, Inc.
  */
 
+var assert = require('assert-plus');
 var buckets = require('../buckets');
-var common = require('../../common');
+var common = require('../common');
 
 function listBucketObjects(req, res, next) {
+    assert.object(req.bucket_data, 'req.bucket_data');
+    assert.uuid(req.bucket_data.id, 'req.bucket_data.id');
 
-    var bucket = req.bucket;
     var log = req.log;
+
     log.debug({
-        bucket: bucket.name
+        bucket: req.bucket_data,
+        params: req.params
     }, 'listBucketObjects: requested');
 
-    var obj = {
-        bucket_data: req.bucket_data
-    };
-
-    /* This function uses an emitter rather than callback pattern for speed. */
-    var mreq = common.bucketList(req.key, req, obj);
+    var mreq = common.listObjects(req, req.bucket_data.id);
 
     mreq.once('error', function onError(err) {
         mreq.removeAllListeners('end');
@@ -44,7 +43,6 @@ function listBucketObjects(req, res, next) {
         res.end();
         next();
     });
-
 }
 
 module.exports = {
