+++ lib/common.js
@@ -841,75 +841,6 @@ function readdir(dir, req, origOpts) {
     return (mreq);
 }
 
-function bucketList(p, req, origOpts) {
-    assert.string(p, 'p');
-    assert.object(req, 'req');
-    assert.object(req.log, 'req.log');
-    assert.object(req.client, 'req.client');
-    assert.object(req.params, 'req.params');
-    assert.optionalObject(origOpts, 'origOpts');
-
-    var client = req.moray;
-    var ee = new EventEmitter();
-    var log = req.log;
-
-    var opts = {
-        path: p,
-        requestId: req.getId(),
-        sorted: req.params.sorted === 'true',
-        prefix: req.params.prefix || '',
-        orderby: req.params.sort === 'mtime' ? 'created' : 'name'
-    };
-
-    if (req.params.limit) {
-        opts.limit = parseInt(req.params.limit, 10);
-        if (isNaN(opts.limit) || opts.limit <= 0) {
-            process.nextTick(function () {
-                ee.emit('error', new InvalidLimitError(req.params.limit));
-            });
-            return (ee);
-        }
-    } else {
-        opts.limit = 0;
-    }
-    assert.number(opts.limit, 'opts.limit');
-    assert.ok(opts.limit >= 0, 'opts.limit >= 0');
-
-    opts = jsprim.mergeObjects(opts, null, origOpts);
-
-    log.debug({
-        path: p
-    }, 'bucketList: entered');
-    var mreq = client.search(opts);
-
-    mreq.on('record', function (r) {
-        if (r.key === req.key) {
-            // XXX does this happen?
-            log.warn({key1: r.key, key2: req.key}, 'keys the same');
-            return;
-        }
-
-        var entry = {
-            name: r.key.split('/').pop(),
-            etag: r.value.etag,
-            size: r.value.contentLength,
-            type: r.value.type,
-            contentType: r.value.contentType,
-            contentMD5: r.value.contentMD5,
-            mtime: new Date(r.value.mtime).toISOString()
-        };
-
-        if (entry.type === 'object')
-            entry.durability = (r.value.sharks || []).length || 0;
-
-        mreq.emit('entry', entry, r);
-    });
-
-    return (mreq);
-}
-
-
-
 ///--- Exports
 
 module.exports = {
@@ -985,8 +916,6 @@ module.exports = {
     loadMetadata: loadMetadata,
 
     readdir: readdir,
-
-    bucketList: bucketList,
 
     addCustomHeaders: addCustomHeaders,
